Converted Colors/ColorProvider to TypeScript
